# scainer/api/openapi.yaml
openapi: 3.0.3
info:
  title: Scainer API
  description: HTTP API веб-сервиса детекции нечестных решений.
  version: 1.0.0
servers:
  - url: /
security:
  - bearerAuth: []
paths:
  /api/auth/login:
    post:
      operationId: postAuthLogin
      summary: Выдать JWT по admin-логину
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешная авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /api/auth/me:
    get:
      operationId: getAuthMe
      summary: Текущий пользователь по JWT
      responses:
        '200':
          description: Идентичность вызывающего
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /api/contests:
    get:
      operationId: getContests
      summary: Список контестов из run.yaml
      responses:
        '200':
          description: Контесты с отметкой последнего импорта
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContestInfo'
        '401':
          $ref: '#/components/responses/ErrorResponse'
    post:
      operationId: postContests
      summary: Зарегистрировать новый контест
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterContestRequest'
      responses:
        '200':
          description: Контест успешно зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContestInfo'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /api/contests/{id}:
    patch:
      operationId: patchContest
      summary: Обновить параллель контеста
      parameters:
        - $ref: '#/components/parameters/ContestID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContestRequest'
      responses:
        '200':
          description: Контест успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContestInfo'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      operationId: deleteContest
      summary: Удалить контест
      parameters:
        - $ref: '#/components/parameters/ContestID'
      responses:
        '204':
          description: Контест успешно удален
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /api/contests/{id}/ejudge-login:
    get:
      operationId: getContestEjudgeLogin
      summary: Учётные данные для браузерного входа в контест ejudge
      parameters:
        - $ref: '#/components/parameters/ContestID'
      responses:
        '200':
          description: base URL, login, password и contest_id для POST в ejudge master
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EjudgeBrowserLogin'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
  /api/contests/{id}/import:
    post:
      operationId: postContestImport
      summary: Поставить дозагрузку посылок в очередь
      description: >
        Только import — детекторы не запускаются. Возвращает jobId сразу; прогресс/статус — GET /api/jobs/{jobId}
      parameters:
        - $ref: '#/components/parameters/ContestID'
      responses:
        '202':
          description: Job поставлен в очередь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /api/contests/{id}/analyze:
    post:
      operationId: postContestAnalyze
      summary: Поставить пересчёт findings в очередь (асинхронно)
      description: >
        Только analyze по уже импортированным посылкам; включает AI-детектор (ManualOnly).
        Прогресс — GET /api/jobs/{jobId}.
      parameters:
        - $ref: '#/components/parameters/ContestID'
      responses:
        '202':
          description: Job поставлен в очередь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /api/contests/{id}/sync:
    post:
      operationId: postContestSync
      summary: Дозагрузить посылки и пересчитать детекторы (Sync)
      description: >
        Import + analyze в одном job; analyze с WithManual (AI). Прогресс — GET /api/jobs/{jobId}.
      parameters:
        - $ref: '#/components/parameters/ContestID'
      responses:
        '202':
          description: Job поставлен в очередь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
...
  /api/contests/{id}/findings:
    get:
      operationId: getContestFindings
      summary: Находки контеста (идемпотентный GET; фактически Analyze)
      parameters:
        - $ref: '#/components/parameters/ContestID'
      responses:
        '200':
          description: Находки (view-модель отчёта)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportData'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /api/contests/{id}/problems:
    get:
      operationId: getContestProblems
      summary: Список задач контеста
      parameters:
        - $ref: '#/components/parameters/ContestID'
      responses:
        '200':
          description: Список задач с информацией об исключении
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProblemInfo'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /api/contests/{id}/excluded-problems:
    put:
      operationId: putExcludedProblems
      summary: Обновить список исключённых задач контеста
      parameters:
        - $ref: '#/components/parameters/ContestID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExcludedProblemsRequest'
      responses:
        '204':
          description: Список исключённых задач успешно обновлен
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /api/contests/{id}/submissions:
    get:
      operationId: getContestSubmissions
      summary: Список посылок контеста из кэша
      parameters:
        - $ref: '#/components/parameters/ContestID'
      responses:
        '200':
          description: Посылки контеста
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubmissionListItem'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /api/contests/{id}/submissions/{submissionId}/comments:
    get:
      operationId: getSubmissionComments
      summary: Live-комментарии посылки (+ sync статуса из ejudge)
      parameters:
        - $ref: '#/components/parameters/ContestID'
        - $ref: '#/components/parameters/SubmissionID'
      responses:
        '200':
          description: Комментарии и актуальный/кэшированный вердикт
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionCommentsResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
        '502':
          $ref: '#/components/responses/ErrorResponse'
    post:
      operationId: postSubmissionComment
      summary: Отправить комментарий к посылке в ejudge
      parameters:
        - $ref: '#/components/parameters/ContestID'
        - $ref: '#/components/parameters/SubmissionID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCommentRequest'
      responses:
        '204':
          description: Комментарий отправлен
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '502':
          $ref: '#/components/responses/ErrorResponse'
  /api/contests/{id}/submissions/{submissionId}/verdict:
    post:
      operationId: postSubmissionVerdict
      summary: Принять/отклонить посылку (comment затем status в ejudge)
      parameters:
        - $ref: '#/components/parameters/ContestID'
        - $ref: '#/components/parameters/SubmissionID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostVerdictRequest'
      responses:
        '204':
          description: Вердикт выставлен
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '502':
          $ref: '#/components/responses/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ContestID:
      name: id
      in: path
      required: true
      schema:
        type: string
    SubmissionID:
      name: submissionId
      in: path
      required: true
      schema:
        type: string
  responses:
    ErrorResponse:
      description: Ошибка
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      required: [access_token, expires_in]
      properties:
        access_token:
          type: string
        expires_in:
          type: integer
          description: TTL токена в секундах
    MeResponse:
      type: object
      required: [username]
      properties:
        username:
          type: string
    EjudgeBrowserLogin:
      type: object
      required: [base_url, login, password, contest_id]
      properties:
        base_url:
          type: string
        login:
          type: string
        password:
          type: string
        contest_id:
          type: integer
    ContestInfo:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
          description: Имя контеста из judge-системы (read-only)
        parallelId:
          type: string
          nullable: true
          description: ID параллели (напр. 3, 10, R, F, X)
        lastImportedAt:
          type: string
          format: date-time
          nullable: true
        computedAt:
          type: string
          format: date-time
          nullable: true
          description: Время последнего успешного analyze
        submissionCount:
          type: integer
          description: Число выгруженных посылок контеста
        problemCount:
          type: integer
          description: Число задач с посылками
    JobResponse:
      type: object
      required: [jobId]
      properties:
        jobId:
          type: string
...
    ReportData:
      type: object
      required: [generated_note, findings, submissions]
      properties:
        generated_note:
          type: string
        findings:
          type: array
          items:
            $ref: '#/components/schemas/FindingView'
        submissions:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/SubmissionView'
    FindingView:
      type: object
      required: [key, subject, score, signals]
      properties:
        key:
          type: string
        subject:
          $ref: '#/components/schemas/SubjectView'
        score:
          type: number
          format: double
        ai:
          type: boolean
        signals:
          type: array
          items:
            $ref: '#/components/schemas/SignalView'
...
    SubmissionView:
      type: object
      required: [participant, problem, lang, submitted_at, verdict, source]
      properties:
        contest:
          type: string
        contest_name:
          type: string
        participant:
          type: string
        problem:
          type: string
        problem_name:
          type: string
        lang:
          type: string
        submitted_at:
          type: string
        verdict:
          type: string
        source:
          type: array
          items:
            type: string
    RegisterContestRequest:
      type: object
      required: [id]
      properties:
        id:
          type: string
          description: ID контеста в judge-системе
        parallelId:
          type: string
          nullable: true
          description: ID параллели (напр. 3, 10, R, F, X)
    UpdateContestRequest:
      type: object
      properties:
        parallelId:
          type: string
          nullable: true
          description: ID параллели (напр. 3, 10, R, F, X)
    ProblemInfo:
      type: object
      required: [id, name, excluded, submissionCount, pendingCount]
      properties:
        id:
          type: string
          description: ID задачи
        name:
          type: string
          description: Имя задачи из посылок
        excluded:
          type: boolean
          description: Исключена ли из анализа
        submissionCount:
          type: integer
          description: Число посылок по этой задаче в store
        pendingCount:
          type: integer
          description: Число посылок со статусом PR (pending review) по этой задаче
    ExcludedProblemsRequest:
      type: object
      required: [problems]
      properties:
        problems:
          type: array
          items:
            type: string
          description: Список ID исключённых задач
    SubmissionListItem:
      type: object
      required: [id, problem, participant, lang, submitted_at, verdict]
      properties:
        id:
          type: string
        problem:
          type: string
        participant:
          type: string
        lang:
          type: string
        submitted_at:
          type: string
          format: date-time
        verdict:
          type: string
    RunComment:
      type: object
      required: [id, from, text, time]
      properties:
        id:
          type: string
        from:
          type: string
          description: author display name (from ejudge view-source)
        subject:
          type: string
        text:
          type: string
        time:
          type: string
          format: date-time
    SubmissionCommentsResponse:
      type: object
      required: [comments, verdict, status_stale, source]
      properties:
        comments:
          type: array
          items:
            $ref: '#/components/schemas/RunComment'
        source:
          type: array
          items:
            type: string
          description: source lines from scainer cache
        verdict:
          type: string
        status_stale:
          type: boolean
        status_error:
          type: string
        comments_error:
          type: string
    PostCommentRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
    PostVerdictRequest:
      type: object
      required: [verdict]
      properties:
        verdict:
          type: string
          enum: [OK, RJ]
        comment:
          type: string